+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 1.0.5
+
+`NEW` 支持设定sln目录
+
+`NEW` 版本号提高到1.0.5这样覆盖掉老版本的emmylua-unity
+
+`NEW` 支持选择框架 xlua / tolua
+
+`NEW` 支持选择注入插件 emmylua / sumneko_lua（但并未实现）
+
 ## 1.0.1
 
 更新readme